terminei o CRUD, ainda falta organizar os responses.
