--- src/Exception/EcommerceDomainException.java
@@ -1,16 +1,0 @@
-package Exception;
-
-public class EcommerceDomainException extends RuntimeException{
-
-    private int statusCode;
-
-    public EcommerceDomainException(String message, int statusCode){
-        super(message);
-        this.statusCode = statusCode;
-    }
-
-
-    public int getStatusCode() {
-        return statusCode;
-    }
-}
