--- src/Exception/IllegalArgumentException.java
@@ -1,8 +1,0 @@
-package Exception;
-
-public class IllegalArgumentException extends EcommerceDomainException{
-
-    public IllegalArgumentException(String message) {
-        super(message, 400);
-    }
-}
