--- src/Exception/ObjectNotFoundException.java
@@ -1,8 +1,0 @@
-package Exception;
-
-public class ObjectNotFoundException extends EcommerceDomainException{
-
-    public ObjectNotFoundException(String message){
-        super(message, 404);
-    }
-}
