--- src/controller/ProductControllerInterface.java
@@ -1,19 +1,0 @@
-package controller;
-
-import entity.Product;
-import response.Response;
-
-import java.util.Set;
-
-public interface ProductControllerInterface {
-
-    public Response<Void> deleteProduct (int id);
-    public Response<Void> updateProduct ( Product product);
-    public Response<Void> addProduct (Product product);
-    public Response<Set<Product>> getProducts ();
-    public Response<Product> getProductById ( int id);
-
-
-
-
-}
