--- src/controller/ProxyFactoryJDK.java
@@ -1,16 +1,0 @@
-package controller;
-
-import java.lang.reflect.Proxy;
-
-public class ProxyFactoryJDK {
-    public static <T> T createProxy(T target, Class<T> interfaceType){
-        return (T) Proxy.newProxyInstance(interfaceType.getClassLoader(),
-                new Class<?>[]{interfaceType},
-                new ProxyInvocationHandler(target)
-                );
-    }
-
-
-
-
-}
