+++ src/main/java/com/fabiovaleiro/Exception/IllegalArgumentException.java
@@ -1,0 +1,8 @@
+package com.fabiovaleiro.Exception;
+
+public class IllegalArgumentException extends EcommerceDomainException{
+
+    public IllegalArgumentException(String message) {
+        super(message, 400);
+    }
+}
