+++ src/main/java/com/fabiovaleiro/Exception/ObjectNotFoundException.java
@@ -1,0 +1,8 @@
+package com.fabiovaleiro.Exception;
+
+public class ObjectNotFoundException extends EcommerceDomainException{
+
+    public ObjectNotFoundException(String message){
+        super(message, 404);
+    }
+}
