+++ src/main/java/com/fabiovaleiro/response/ResponseInterface.java
@@ -1,0 +1,14 @@
+package com.fabiovaleiro.response;
+
+public interface ResponseInterface<T> {
+
+    public T getData();
+
+    public void setData(T data);
+
+    public ErrorDetails getError();
+
+    public void setError(ErrorDetails error);
+
+
+}
