+++ src/main/resources/db-config.properties
@@ -1,0 +1,3 @@
+db.url=jdbc:mysql://localhost:3306/ecommerce
+db.username=root
+db.password=Sucesso@777
