--- src/response/ResponseInterface.java
@@ -1,14 +1,0 @@
-package response;
-
-public interface ResponseInterface<T> {
-
-    public T getData();
-
-    public void setData(T data);
-
-    public ErrorDetails getError();
-
-    public void setError(ErrorDetails error);
-
-
-}
